Rename OperationCollection, related files, and rename conflicting types
