+++ LocalPackages/DataBrokerProtection/Tests/DataBrokerProtectionTests/DataBrokerOperationActionTests.swift
@@ -40,7 +40,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
         let emailConfirmationAction = EmailConfirmationAction(id: "", actionType: .emailConfirmation, pollingTime: 1, dataSource: nil)
         let step = Step(type: .optOut, actions: [emailConfirmationAction])
         let extractedProfile = ExtractedProfile(email: "test@duck.com")
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -65,7 +65,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
         let emailConfirmationAction = EmailConfirmationAction(id: "", actionType: .emailConfirmation, pollingTime: 1, dataSource: nil)
         let step = Step(type: .optOut, actions: [emailConfirmationAction])
         let noEmailExtractedProfile = ExtractedProfile()
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -97,7 +97,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
         let step = Step(type: .optOut, actions: [emailConfirmationAction])
         let extractedProfile = ExtractedProfile(email: "test@duck.com")
         emailService.shouldThrow = true
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -127,7 +127,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenActionNeedsEmail_thenExtractedProfileEmailIsSet() async {
         let fillFormAction = FillFormAction(id: "1", actionType: .fillForm, selector: "#test", elements: [.init(type: "email", selector: "#email", parent: nil)], dataSource: nil)
         let step = Step(type: .optOut, actions: [fillFormAction])
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -150,7 +150,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenGetEmailServiceFails_thenOperationThrows() async {
         let fillFormAction = FillFormAction(id: "1", actionType: .fillForm, selector: "#test", elements: [.init(type: "email", selector: "#email", parent: nil)], dataSource: nil)
         let step = Step(type: .optOut, actions: [fillFormAction])
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -178,7 +178,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     }
 
     func testWhenClickActionSucceeds_thenWeWaitForWebViewToLoad() async {
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -199,7 +199,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     }
 
     func testWhenAnActionThatIsNotClickSucceeds_thenWeDoNotWaitForWebViewToLoad() async {
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -221,7 +221,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenSolveCaptchaActionIsRun_thenCaptchaIsResolved() async {
         let solveCaptchaAction = SolveCaptchaAction(id: "1", actionType: .solveCaptcha, selector: "g-captcha", dataSource: nil)
         let step = Step(type: .optOut, actions: [solveCaptchaAction])
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -244,7 +244,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenSolveCapchaActionFailsToSubmitDataToTheBackend_thenOperationFails() async {
         let solveCaptchaAction = SolveCaptchaAction(id: "1", actionType: .solveCaptcha, selector: "g-captcha", dataSource: nil)
         let step = Step(type: .optOut, actions: [solveCaptchaAction])
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(with: [step]),
@@ -274,7 +274,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenCaptchaInformationIsReturned_thenWeSubmitItTotTheBackend() async {
         let getCaptchaResponse = GetCaptchaInfoResponse(siteKey: "siteKey", url: "url", type: "recaptcha")
         let step = Step(type: .optOut, actions: [Action]())
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -297,7 +297,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenCaptchaInformationFailsToBeSubmitted_thenTheOperationFails() async {
         let getCaptchaResponse = GetCaptchaInfoResponse(siteKey: "siteKey", url: "url", type: "recaptcha")
         let step = Step(type: .optOut, actions: [Action]())
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -321,7 +321,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
 
     func testWhenRunningActionWithoutExtractedProfile_thenExecuteIsCalledWithProfileData() async {
         let expectationAction = ExpectationAction(id: "1", actionType: .expectation, expectations: [Item](), dataSource: nil)
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -340,7 +340,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     }
 
     func testWhenLoadURLDelegateIsCalled_thenCorrectMethodIsExecutedOnWebViewHandler() async {
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -361,7 +361,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenGetCaptchaActionRuns_thenStageIsSetToCaptchaParse() async {
         let mockStageCalculator = MockStageDurationCalculator()
         let captchaAction = GetCaptchaInfoAction(id: "1", actionType: .getCaptchaInfo, selector: "captcha", dataSource: nil)
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -381,7 +381,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenClickActionRuns_thenStageIsSetToSubmit() async {
         let mockStageCalculator = MockStageDurationCalculator()
         let clickAction = ClickAction(id: "1", actionType: .click, elements: [PageElement](), dataSource: nil)
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -401,7 +401,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenExpectationActionRuns_thenStageIsSetToSubmit() async {
         let mockStageCalculator = MockStageDurationCalculator()
         let expectationAction = ExpectationAction(id: "1", actionType: .expectation, expectations: [Item](), dataSource: nil)
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -421,7 +421,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
     func testWhenFillFormActionRuns_thenStageIsSetToFillForm() async {
         let mockStageCalculator = MockStageDurationCalculator()
         let fillFormAction = FillFormAction(id: "1", actionType: .fillForm, selector: "", elements: [PageElement](), dataSource: nil)
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(),
@@ -440,7 +440,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
 
     func testWhenLoadUrlOnSpokeo_thenSetCookiesIsCalled() async {
         let mockCookieHandler = MockCookieHandler()
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(url: "spokeo.com"),
@@ -462,7 +462,7 @@ final class DataBrokerOperationActionTests: XCTestCase {
 
     func testWhenLoadUrlOnOtherBroker_thenSetCookiesIsNotCalled() async {
         let mockCookieHandler = MockCookieHandler()
-        let sut = OptOutOperation(
+        let sut = OptOutTask(
             privacyConfig: PrivacyConfigurationManagingMock(),
             prefs: ContentScopeProperties.mock,
             query: BrokerProfileQueryData.mock(url: "verecor.com"),
